+++ e2e/cypress/elements/actions-selectors.js
@@ -1,0 +1,4 @@
+export const ACTION_SELECTORS = {
+  delete: "[data-test='btnDeleteTest']",
+  confirmModal: "[data-test='btnConfirmModalTest']",
+}