+++ e2e/cypress/elements/breadcrumb-selectors.js
@@ -1,0 +1,3 @@
+export const BREADCRUMB_SELECTORS = {
+  prefixItem: "breadcrumbItemTest"
+}