+++ e2e/cypress/elements/panels-selectors.js
@@ -1,0 +1,18 @@
+export const ITEMS_PANELS_SELECTORS = {
+  table: "[data-test='itemPanelTableTest']",
+  prefixItem: "itemTest"
+}
+
+export const CONTEXT_TOOLBAR_SELECTORS = {
+  btnAddType: "[data-test='itemAddTypeTest']",
+  btnAddFolder: "[data-test='dropdownItemTest-folder']",
+  btnAddItem: "[data-test='dropdownItemTest-item']",
+  btnAddGMI: "[data-test='dropdownItemTest-gmi']",
+  selectFilteType: "[data-test='selectFilterTypeTest']",
+  inputFilter: "[data-test='inputFilterTest']",
+  btnInputFilter: "[data-test='btnInputFilterTest']"
+}
+
+export const TABS_PANEL_SELECTOS = {
+  prefixTabs: "tabTest"
+}