+++ e2e/cypress/elements/user-form-selectors.js
@@ -1,6 +1,0 @@
-export const USER_FORM_SELECTORS = {
-  username: "[data-test='usernameTestInput']",
-  password: "[data-test='passwordTestInput']",
-  email: "[data-test='emailTestInput']",
-  name: "[data-test='nameTestInput']",
-}