+++ src/guillo-gmi/components/fields/editComponent.js
@@ -1,0 +1,101 @@
+import React from 'react'
+import { Textarea } from '../input/textarea'
+import { Checkbox } from '../input/checkbox'
+import { FileUpload } from '../input/upload'
+import { Select } from '../input/select'
+import { Input } from '../input/input'
+import { InputList } from '../input/input_list'
+import { get } from '../../lib/utils'
+
+export const EditComponent = React.forwardRef(
+  ({ schema, val, setValue, dataTest, className, ...rest }, ref) => {
+    if (schema?.widget === 'textarea' || schema?.widget === 'richtext') {
+      return (
+        <Textarea
+          value={val || ''}
+          className={className}
+          onChange={(ev) => setValue(ev)}
+          ref={ref}
+          dataTest={dataTest}
+          {...rest}
+        />
+      )
+    } else if (schema?.type === 'boolean') {
+      return (
+        <Checkbox
+          value={!!val}
+          className={className}
+          onChange={(ev) => setValue(ev)}
+          ref={ref}
+          dataTest={dataTest}
+          {...rest}
+        />
+      )
+    } else if (schema?.type === 'array') {
+      return (
+        <InputList
+          value={val || []}
+          className={className}
+          onChange={(ev) => setValue(ev)}
+          ref={ref}
+          dataTest={dataTest}
+          {...rest}
+        />
+      )
+    } else if (schema?.widget === 'file') {
+      return (
+        <FileUpload
+          onChange={(ev) => setValue(ev)}
+          label={get(val, 'filename', null)}
+          dataTest={dataTest}
+          {...rest}
+        />
+      )
+    } else if (schema?.widget === 'select') {
+      return (
+        <Select
+          value={val || ''}
+          className={className}
+          appendDefault
+          classWrap="is-fullwidth"
+          dataTest={dataTest}
+          {...rest}
+          options={schema?.vocabulary.map((item) => {
+            return {
+              text: item,
+              value: item,
+            }
+          })}
+          onChange={(ev) => {
+            const selectValue = get(ev, 'target.value', '')
+            return setValue(selectValue)
+          }}
+        />
+      )
+    }
+    const getInputType = () => {
+      switch (schema?.type) {
+        case 'integer':
+          return 'number'
+        case 'datetime':
+          return 'datetime-local'
+        default:
+          return 'text'
+      }
+    }
+    return (
+      <Input
+        value={val || ''}
+        className={className}
+        dataTest={dataTest}
+        onChange={(ev) => setValue(ev)}
+        ref={ref}
+        type={getInputType()}
+        {...rest}
+      />
+    )
+  }
+)
+
+EditComponent.displayName = 'EditComponent'
+export default EditComponent