+++ src/guillo-gmi/components/input/checkbox.js
@@ -4,6 +4,7 @@ import { classnames } from '../../lib/helpers'
 export const Checkbox = ({
   id,
   className,
+  classNameInput,
   loading,
   disabled,
   indeterminate = false,
@@ -14,6 +15,7 @@ export const Checkbox = ({
   children,
   placeholder,
   onChange,
+  dataTest,
   ...rest
 }) => {
   const inputRef = useRef(null)
@@ -38,9 +40,10 @@ export const Checkbox = ({
           disabled={disabled || loading}
           id={id}
           type="checkbox"
-          className="checkbox"
+          className={classnames(['checkbox', classNameInput])}
           checked={state}
           onChange={updateState}
+          data-test={dataTest}
           {...rest}
         />
         {children || placeholder}