+++ src/guillo-gmi/components/input/select.js
@@ -1,7 +1,7 @@
-import React from 'react'
+import React, { useState } from 'react'
 import PropTypes from 'prop-types'
-
-import { classnames } from '../../lib/helpers'
+import ErrorZone from '../error_zone'
+import { classnames, generateUID } from '../../lib/helpers'
 
 // @ TODO implement hasErrors
 
@@ -11,7 +11,10 @@ export const Select = React.forwardRef(
     {
       options,
       error,
+      errorZoneClassName,
       size = 1,
+      placeholder,
+      id,
       className = '',
       classWrap = '',
       multiple = false,
@@ -20,10 +23,13 @@ export const Select = React.forwardRef(
       onChange,
       appendDefault = false,
       style = {},
+      dataTest,
       ...rest
     },
     ref
   ) => {
+    const [uid] = useState(generateUID('select'))
+
     const onUpdate = (ev) => {
       if (ev.target.value === '') {
         onChange({ target: { value: undefined } })
@@ -35,27 +41,46 @@ export const Select = React.forwardRef(
     if (appendDefault) {
       options = [{ text: 'Choose..', value: '' }].concat(options)
     }
+    const statusClasses = error ? 'is-danger' : ''
 
-    const cssWrap = ['select', multiple ? 'is-multiple' : '', classWrap]
+    const cssWrap = [
+      'select',
+      statusClasses,
+      multiple ? 'is-multiple' : '',
+      classWrap,
+    ]
 
     return (
-      <div className={classnames(cssWrap)}>
-        <select
-          className={classnames(['', className])}
-          size={size}
-          multiple={multiple}
-          disabled={loading || rest.disabled}
-          onChange={onUpdate}
-          {...rest}
-          ref={ref}
-          style={style}
-        >
-          {options.map(({ text, ...rest }, index) => (
-            <option key={index.toString()} {...rest}>
-              {text}
-            </option>
-          ))}
-        </select>
+      <div className="field">
+        {id && placeholder ? (
+          <label className="label" htmlFor={id}>
+            {placeholder}
+          </label>
+        ) : null}
+        <div className={classnames(cssWrap)}>
+          <select
+            className={classnames(['', className])}
+            size={size}
+            multiple={multiple}
+            disabled={loading || rest.disabled}
+            onChange={onUpdate}
+            {...rest}
+            ref={ref}
+            style={style}
+            data-test={dataTest}
+          >
+            {options.map(({ text, ...rest }, index) => (
+              <option key={index.toString()} {...rest}>
+                {text}
+              </option>
+            ))}
+          </select>
+        </div>
+        {error && (
+          <ErrorZone className={errorZoneClassName} id={uid}>
+            {error ? error : ''}
+          </ErrorZone>
+        )}
       </div>
     )
   }