+++ src/guillo-gmi/components/panel/index.js
@@ -1,6 +1,7 @@
 // @create-index
 
 export { default as addons } from './addons.js'
+export { default as actions } from './actions.js'
 export { default as docs } from './docs.js'
 export { default as item } from './item.js'
 export { default as items } from './items.js'