+++ src/guillo-gmi/components/ui/delete.js
@@ -1,8 +1,13 @@
 import React from 'react'
 
-export function Delete({ onClick, children }) {
+export function Delete({ onClick, className, children }) {
   return (
-    <button type="button" onClick={onClick} className="delete">
+    <button
+      type="button"
+      onClick={onClick}
+      className={`delete ${className}`}
+      data-test="btnDeleteTest"
+    >
       {children}
     </button>
   )